Only polyfill older phpunit methods when required
